+++ src/common/timer.cpp
@@ -17,5 +17,5 @@ float CTimer::ElapsedTime() const
 {
     high_resolution_clock::time_point now = high_resolution_clock::now();
     duration<double> time_span = duration_cast<duration<double>>(now - startTime);
-    return time_span.count();
+    return static_cast<float>(time_span.count());
 }
